chore: Move AIParametersParser to n8n-workflow (no-changelog) (#12671)
