+++ packages/workflow/src/index.ts
@@ -13,6 +13,7 @@ export * from './Interfaces';
 export * from './MessageEventBus';
 export * from './ExecutionStatus';
 export * from './Expression';
+export * from './FromAIParseUtils';
 export * from './NodeHelpers';
 export * from './Workflow';
 export * from './WorkflowDataProxy';
